+++ src/pa2/utils/BinaryUtils.java
@@ -2,10 +2,10 @@ package pa2.utils;
 
 public class BinaryUtils {
 
-    public static int parseBinary(String num){
-        int mod = 1;
-        int total = 0;
-        int temp;
+    public static long parseBinary(String num){
+        long mod = 1;
+        long total = 0;
+        long temp;
         for(int i = 0; i < num.length();i++){
             temp = Integer.parseInt(num.substring(num.length()-i-1,num.length()-i));
             total += temp*mod;
